requirements.txt
# LLM & Agent
openai>=1.0.0
langchain>=0.2.0
langgraph>=0.1.0

# HTTP & API
httpx>=0.27.0
aiohttp>=3.9.0

# Text similarity
sentence-transformers>=2.2.0
rapidfuzz>=3.0.0

# Google Scholar fallback (direct HTTP scraping)
beautifulsoup4>=4.12.0

# Web framework
fastapi>=0.110.0
uvicorn>=0.29.0
sse-starlette>=2.0.0
gradio>=4.0.0

# Data processing
pydantic>=2.0.0
pyyaml>=6.0.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
